Adding the middleware pipleline behavior with Exception handling, Logging(SeriLog) and Authentication(JWT token)

+++ KakaoTicket.TicketManagement.Application/Features/Events/Commands/DeleteEvent/DeleteEventCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using KakaoTicket.TicketManagement.Application.Contracts.Persistence;
+using KakaoTicket.TicketManagement.Application.Exceptions;
 using KakaoTicket.TicketManagement.Domain.Entities;
 using MediatR;
 using System.Threading;
@@ -21,6 +22,11 @@ namespace KakaoTicket.TicketManagement.Application.Features.Events.Commands.Dele
         public async Task<Unit> Handle(DeleteEventCommand request, CancellationToken cancellationToken)
         {
             var eventToDelete = await _eventRepository.GetByIdAsync(request.EventId);
+
+            if (eventToDelete == null)
+            {
+                throw new NotFoundException(nameof(Event), request.EventId);
+            }
 
             await _eventRepository.DeleteAsync(eventToDelete);
